pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ai-talent-scout"
version = "2.0.0"
description = "AI-powered talent scouting & engagement agent — discovers, engages, and ranks candidates using sentence-transformers, FAISS, and conversational simulation."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "sentence-transformers>=3.0.0",
    "torch>=2.6.0",
    "transformers>=4.41.0",
    "huggingface-hub>=0.23.0",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",
    "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",       # for TestClient async support
    "ruff>=0.4",         # linting & formatting
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.setuptools.packages.find]
include = ["app*"]
